# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Credit Card Fraud Detection API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /cards:
    get:
      operationId: listCards
      tags: [cards]
      summary: List cards for the authenticated user
      responses:
        "200":
          description: Card list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"

    post:
      operationId: createCard
      tags: [cards]
      summary: Add a new card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardInput"
      responses:
        "201":
          description: Card created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
  /cards/{id}:
    get:
      operationId: getCard
      tags: [cards]
      summary: Get card details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Card detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
  /cards/{id}/block:
    patch:
      operationId: blockCard
      tags: [cards]
      summary: Block or unblock a card
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardBlockInput"
      responses:
        "200":
          description: Card status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
...
  /fraud-cases:
    get:
      operationId: listFraudCases
      tags: [admin]
      summary: List fraud cases
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Fraud cases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudCaseListResponse"
  /fraud-cases/{id}:
    get:
      operationId: getFraudCase
      tags: [admin]
      summary: Get fraud case
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fraud case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudCase"
    patch:
      operationId: updateFraudCase
      tags: [admin]
      summary: Update fraud case
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                priority:
                  type: string
                resolution:
                  type: string
                assignedTo:
                  type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudCase"
...
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
      responses:
        "200":
          description: Reset email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, token, newPassword]
              properties:
                email:
                  type: string
                token:
                  type: string
                newPassword:
                  type: string
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, name, role, status, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [active, suspended]
        createdAt:
          type: string
...
    Card:
      type: object
      required: [id, userId, last4, brand, expiryMonth, expiryYear, isBlocked, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        last4:
          type: string
        brand:
          type: string
          enum: [visa, mastercard, amex, discover]
        expiryMonth:
          type: integer
        expiryYear:
          type: integer
        isBlocked:
          type: boolean
        blockReason:
          type: ["string", "null"]
        createdAt:
          type: string
    CardInput:
      type: object
      required: [last4, brand, expiryMonth, expiryYear]
      properties:
        last4:
          type: string
        brand:
          type: string
          enum: [visa, mastercard, amex, discover]
        expiryMonth:
          type: integer
        expiryYear:
          type: integer
    CardBlockInput:
      type: object
      required: [blocked]
      properties:
        blocked:
          type: boolean
        reason:
          type: string
...
    SuccessMessage:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    FraudCase:
      type: object
      required: [id, caseNumber, userId, status, priority, title, createdAt, updatedAt]
      properties:
        id:
          type: integer
        caseNumber:
          type: string
        userId:
          type: integer
        transactionId:
          type: ["integer", "null"]
        assignedTo:
          type: ["integer", "null"]
        status:
          type: string
        priority:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        resolution:
          type: ["string", "null"]
        riskScore:
          type: ["number", "null"]
        amountInvolved:
          type: ["number", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        closedAt:
          type: ["string", "null"]
    FraudCaseListResponse:
      type: object
      required: [cases, total, page, limit]
      properties:
        cases:
          type: array
          items:
            $ref: "#/components/schemas/FraudCase"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...