# lib/api-zod/src/generated/types/fraudCaseListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Credit Card Fraud Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { FraudCase } from './fraudCase';

export interface FraudCaseListResponse {
  cases: FraudCase[];
  total: number;
  page: number;
  limit: number;
}